Add workaround to settings navigation item
The problem here is that setting the navigationItem's title on viewDidLoad doesn't work for SettingsViewController because this only gets called
when the user taps on the UITabBar. The workaround in place is to set the title via Interface Builder, but the problem here is that this string
isn't going to be localized. Proper fix would be to set the navigationItem's title before the whole rootViewController is displayed, ugly and
not neccesary for now.